Extract Downloadable concern from Mod
